# interop-resolution-service/docs/openapi/interop-resolution-service-openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Interop Resolution Service",
    "description": "Onyx's interop lookups. Member resolution, POST /v1/interop/resolve: resolves an EMR-supplied member ID against the plan's member records, reports whether coverage is active on the date of service, and returns the ID as resolved and in every UM vendor's format, each with the payer for that vendor's request.",
    "version": "v1"
  },
  "tags": [
    {
      "name": "Member resolution",
      "description": "Resolve an EMR-supplied member id against the plan's member records for a date of service"
    }
  ],
  "paths": {
    "/v1/interop/resolve": {
      "post": {
        "tags": [
          "Member resolution"
        ],
        "summary": "Resolve a member id for a date of service",
        "description": "Finds the member behind the id the EMR typed, says whether coverage is active on the date of service (or on every day of a period of service), and returns the id as resolved and in every configured UM vendor's format (memberId.forVendors), each entry with the payer for that vendor's request (payerId, payerName). When several members on the plan match and no dateOfBirth settles it, the outcome is AMBIGUOUS and the message says what to resend; the other members are never listed. clientId, clientType (EXT or INT), requestId and memberId are required; unknown properties are ignored. A sent date is the date evaluated; an unusable date is a 400. POST because the member id is PHI and must not appear in URLs; the operation is a pure read and may be repeated.",
        "operationId": "resolve",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MemberResolutionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "outcome ACTIVE | INACTIVE | NOT_FOUND | AMBIGUOUS; memberId.forVendors, each vendor with its payer, present for ACTIVE and INACTIVE",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MemberResolutionResponse"
                }
              }
            }
          },
          "400": {
            "description": "INVALID_REQUEST for a missing or unusable clientId, clientType or requestId, a missing memberId, an unusable date or a body that is not JSON; MEMBER_LOOKUP_REJECTED when the member lookup answered 400 (details[0].code HTTP_400, the lookup's text in the message)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "DOB_MISMATCH: the date of birth matches no record for this id",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "502": {
            "description": "MEMBER_LOOKUP_ERROR or MEMBER_LOOKUP_INVALID_RESPONSE (the member lookup); MEMBER_PLAN_ERROR or MEMBER_PLAN_INVALID_RESPONSE (the member plan lookup)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "MEMBER_LOOKUP_UNAVAILABLE or MEMBER_PLAN_UNAVAILABLE; retry after the Retry-After seconds",
            "headers": {
              "Retry-After": {
                "description": "Seconds to wait before retrying",
                "style": "simple",
                "schema": {
                  "type": "integer",
                  "example": 10
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "INTERNAL_ERROR: the service itself failed; retry once later, then alert the service owners with the correlationId and requestId",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "MemberResolutionRequest": {
        "type": "object",
        "description": "Member resolution request: clientId, clientType, requestId and memberId are required; unknown properties are ignored; a sent date must be usable",
        "properties": {
          "clientId": {
            "type": "string",
            "description": "The calling system, as registered with the plan: letters, digits, '.', '_' or '-', at most 50 characters",
            "example": "ONYX",
            "maxLength": 50,
            "pattern": "^[A-Za-z0-9._-]{1,50}$"
          },
          "clientType": {
            "type": "string",
            "description": "EXT for a caller outside the plan, INT for an internal one",
            "enum": [
              "EXT",
              "INT"
            ],
            "example": "EXT"
          },
          "requestId": {
            "type": "string",
            "description": "The caller's id for this call, unique per request; echoed in the response and in every error response. Letters, digits, '.', '_', ':' or '-', at most 64 characters",
            "example": "3f6c2a9e-8b1d-4e7a-9c5f-2d4b6a8e0c13",
            "maxLength": 64,
            "pattern": "^[A-Za-z0-9._:-]{1,64}$"
          },
          "memberId": {
            "type": "string",
            "description": "The member id exactly as the EMR supplied it. Looked up unchanged (surrounding whitespace removed); no shape, length or character check. The example is a masked pattern: a THP Medicare card number is a letter and 8 digits",
            "example": "S########"
          },
          "dateOfService": {
            "type": "string",
            "description": "Date of service, yyyy-MM-dd; the first date when the service covers a period. Defaults to today when omitted (dateOfServiceDefaulted is true). A sent value is the date evaluated: any future date is judged on the coverage on record; not a real date, or more than 10 years back, is a 400.",
            "example": "2026-10-15"
          },
          "dateOfServiceEnd": {
            "type": "string",
            "description": "Optional last date of service, yyyy-MM-dd, when the service covers a period: coverage must hold on every day from dateOfService to it. Not before dateOfService; needs dateOfService.",
            "example": "2026-10-20"
          },
          "dateOfBirth": {
            "type": "string",
            "description": "Optional patient date of birth, yyyy-MM-dd, used only to verify the member or to tell apart the members on a plan who share the id. Never logged, never echoed.",
            "example": "1950-03-15"
          }
        },
        "required": [
          "clientId",
          "clientType",
          "memberId",
          "requestId"
        ]
      },
      "Coverage": {
        "type": "object",
        "description": "The coverage flag and the continuous coverage period that covers the (first) date of service. The dates are present when active, and when inactive only because the period ends before the last date asked about; otherwise absent, and message says why",
        "properties": {
          "coverageId": {
            "type": "string",
            "description": "An id for the coverage period: the resolved id reduced to letters and digits, then the effective and end dates as yyyyMMdd, run together with no separator (39991231 for an open-ended period). Present whenever the dates are; the same member with the same coverage always gets the same value"
          },
          "active": {
            "type": "boolean",
            "description": "The flag: true when coverage holds on the date of service, or on every day of the period of service"
          },
          "effectiveDate": {
            "type": "string",
            "format": "date",
            "description": "First day of the coverage period"
          },
          "endDate": {
            "type": "string",
            "format": "date",
            "description": "Last day of the coverage period; 3999-12-31 for an open-ended period"
          }
        },
        "required": [
          "active"
        ]
      },
      "MemberId": {
        "type": "object",
        "description": "The member id as received, as resolved, and in every vendor's format",
        "properties": {
          "received": {
            "type": "string",
            "description": "The member id as sent, surrounding whitespace removed"
          },
          "resolved": {
            "type": "string",
            "description": "The id the lookup resolved to, as the plan stores it; for a converted member a different number from the one received. ACTIVE and INACTIVE only"
          },
          "forVendors": {
            "type": "array",
            "description": "The resolved id in every configured vendor's format, each with the payer for that vendor's request; one entry per vendor, sorted by vendor code. ACTIVE and INACTIVE only",
            "items": {
              "$ref": "#/components/schemas/VendorMemberId"
            }
          }
        },
        "required": [
          "received"
        ]
      },
      "MemberResolutionResponse": {
        "type": "object",
        "description": "The answer: the outcome Onyx branches on, one sentence for a person, the member id as received, as resolved and in every vendor's format with that vendor's payer, the line of business, the date evaluated, the coverage flag with its period, the caller's requestId and a traceId for support. A field that does not apply is absent, never null",
        "properties": {
          "outcome": {
            "type": "string",
            "description": "ACTIVE: member found, coverage active on the date of service (or on every day of the period). INACTIVE: member found, not covered on that date or not on every day of the period; message says why. NOT_FOUND: no member for this id on that date. AMBIGUOUS: several members match and nothing sent tells them apart; message says what to resend",
            "enum": [
              "ACTIVE",
              "INACTIVE",
              "NOT_FOUND",
              "AMBIGUOUS"
            ]
          },
          "message": {
            "type": "string",
            "description": "One sentence for a person reading the answer. Branch on outcome, not on this text"
          },
          "memberId": {
            "$ref": "#/components/schemas/MemberId"
          },
          "lineOfBusiness": {
            "type": "string",
            "description": "The line of business Onyx routes on, from the plan in effect on the date of service: SCO, MA-HMO, MA-PPO, D-SNP, MA-TOGETHER, RI-TOGETHER or MA-QHP; a member no rule covers keeps the member record's category (MCR, PP, COM). ACTIVE and INACTIVE only",
            "example": "MA-HMO"
          },
          "dateOfService": {
            "type": "string",
            "format": "date",
            "description": "The date evaluated; the first date when a period was asked about"
          },
          "dateOfServiceEnd": {
            "type": "string",
            "format": "date",
            "description": "The last date evaluated; present only when the request asked about a period"
          },
          "dateOfServiceDefaulted": {
            "type": "boolean",
            "description": "true when no dateOfService was sent and today was used"
          },
          "coverage": {
            "$ref": "#/components/schemas/Coverage",
            "description": "The coverage flag and the coverage period behind it. ACTIVE and INACTIVE only"
          },
          "requestId": {
            "type": "string",
            "description": "The caller's requestId, echoed"
          },
          "traceId": {
            "type": "string",
            "description": "The id of the member lookup behind this answer; quote it to support"
          },
          "sourceMessage": {
            "$ref": "#/components/schemas/SourceMessage",
            "description": "What the member lookup said when it had no member for the id. NOT_FOUND only, and only when the lookup sent a message"
          }
        },
        "required": [
          "dateOfService",
          "dateOfServiceDefaulted",
          "memberId",
          "message",
          "outcome",
          "requestId",
          "traceId"
        ]
      },
      "SourceMessage": {
        "type": "object",
        "description": "What the member lookup said when it had no member for the id, forwarded as it came",
        "properties": {
          "type": {
            "type": "string",
            "description": "The message type as the lookup sent it",
            "example": "ERROR"
          },
          "status": {
            "type": "string",
            "description": "The status as the lookup sent it",
            "example": "404"
          },
          "code": {
            "type": "string",
            "description": "The message code as the lookup sent it",
            "example": "MEMBER_NOT_FOUND"
          },
          "text": {
            "type": "string",
            "description": "The message text as the lookup sent it"
          }
        }
      },
      "VendorMemberId": {
        "type": "object",
        "description": "Everything Onyx puts on one vendor's request",
        "properties": {
          "vendor": {
            "type": "string",
            "description": "The vendor code",
            "enum": [
              "CARELON",
              "EVICORE",
              "EVOLENT",
              "MHK",
              "ONYX",
              "OPTUM"
            ],
            "example": "EVICORE"
          },
          "memberId": {
            "type": "string",
            "description": "The resolved id in this vendor's format: the member id to put on this vendor's request"
          },
          "payerId": {
            "type": "string",
            "description": "The payer id to put on this vendor's request",
            "example": "TUFTS"
          },
          "payerName": {
            "type": "string",
            "description": "The payer name that goes with payerId",
            "example": "TUFTS"
          }
        }
      },
      "ApiErrorResponse": {
        "type": "object",
        "description": "The shape of every non-200 answer",
        "properties": {
          "error": {
            "$ref": "#/components/schemas/Error"
          },
          "correlationId": {
            "type": "string",
            "description": "The X-Correlation-Id of the call, echoed or generated"
          },
          "requestId": {
            "type": "string",
            "description": "The caller's requestId, when the request carried a usable one"
          },
          "traceId": {
            "type": "string",
            "description": "The id of the member lookup behind the answer, present once the lookup was made"
          }
        },
        "required": [
          "correlationId",
          "error"
        ]
      },
      "Error": {
        "type": "object",
        "description": "The error: a stable code to branch on, a message for a person, and one detail per problem",
        "properties": {
          "code": {
            "type": "string",
            "description": "INVALID_REQUEST, MEMBER_LOOKUP_REJECTED, DOB_MISMATCH, MEMBER_LOOKUP_ERROR, MEMBER_LOOKUP_INVALID_RESPONSE, MEMBER_LOOKUP_UNAVAILABLE, MEMBER_PLAN_ERROR, MEMBER_PLAN_INVALID_RESPONSE, MEMBER_PLAN_UNAVAILABLE or INTERNAL_ERROR",
            "example": "INVALID_REQUEST"
          },
          "message": {
            "type": "string",
            "description": "One sentence for a person"
          },
          "details": {
            "type": "array",
            "description": "One entry per problem found; every problem with a request is reported together",
            "items": {
              "$ref": "#/components/schemas/ErrorDetail"
            }
          }
        },
        "required": [
          "code",
          "message"
        ]
      },
      "ErrorDetail": {
        "type": "object",
        "description": "One problem inside an error response",
        "properties": {
          "field": {
            "type": "string",
            "description": "The request field the problem is about, when there is one",
            "example": "dateOfService"
          },
          "code": {
            "type": "string",
            "description": "The detail code that refines error.code, for example MEMBER_ID_MISSING, DATE_OF_SERVICE_INVALID, HTTP_400, READ_TIMEOUT",
            "example": "DATE_OF_SERVICE_INVALID"
          },
          "message": {
            "type": "string",
            "description": "One sentence for a person"
          }
        },
        "required": [
          "code",
          "message"
        ]
      }
    }
  }
}